+++ src/main/java/com/Aster/FloristApplication.java
@@ -1,0 +1,15 @@
+package com.Aster;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class FloristApplication {
+
+	public static void main(String[] args) {
+
+		SpringApplication.run(FloristApplication.class, args);
+		System.out.println("Welcome to Aster");
+	}
+
+}
