+++ src/main/java/com/Aster/controller/FloristController.java
@@ -1,0 +1,4 @@
+package com.Aster.controller;
+
+public class FloristController {
+}
