+++ src/main/java/com/Aster/controller/ProductController.java
@@ -1,0 +1,4 @@
+package com.Aster.controller;
+
+public class ProductController {
+}
