+++ src/main/java/com/Aster/domain/Florist.java
@@ -1,0 +1,4 @@
+package com.Aster.domain;
+
+public class Florist {
+}
