+++ src/main/java/com/Aster/domain/User.java
@@ -1,0 +1,4 @@
+package com.Aster.domain;
+
+public class User {
+}
