--- src/main/java/com/Aster/florist/FloristApplication.java
@@ -1,15 +1,0 @@
-package com.Aster.florist;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class FloristApplication {
-
-	public static void main(String[] args) {
-
-		SpringApplication.run(FloristApplication.class, args);
-		System.out.println("Welcome to Aster");
-	}
-
-}
