+++ src/test/java/com/Aster/FloristApplicationTests.java
@@ -1,0 +1,13 @@
+package com.Aster;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class FloristApplicationTests {
+
+	@Test
+	void contextLoads() {
+	}
+
+}
