--- src/test/java/com/Aster/florist/FloristApplicationTests.java
@@ -1,13 +1,0 @@
-package com.Aster.florist;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class FloristApplicationTests {
-
-	@Test
-	void contextLoads() {
-	}
-
-}
